move add_normalize_unsafe
